refactor: move commit-batch validation to ungated repository home
The six commit-batch validation fns (reject_duplicate_streams,
reject_duplicate_outbox_messages, validate_entity_id_matches_identity,
validate_prepared_appends, validate_supported_event_codec,
validate_snapshot_identity) were duplicated byte-for-byte between
src/sqlx_repo/mod.rs and src/hashmap_repo/repository.rs. They have no
sqlx dependency; only the cfg gate forced the copy.

Move them to a single ungated home, src/repository/validation.rs, since
they validate repository-level types (StreamWrite, PreparedEventAppend,
OutboxMessage, CommitBatch). All three backends (hashmap, sqlite,
postgres) now import the one copy via crate::repository, so the
definition of a valid CommitBatch cannot drift between backends.

No behavior change: the moved fns are byte-identical to the originals;
the error taxonomy work from #83 in sqlx_repo is untouched.

Implements [[tasks/shared-helpers-dedup-batch]]
