+++ src/hashmap_repo/repository.rs
@@ -7,9 +7,7 @@ use std::collections::{HashMap, HashSet};
 use std::future::Future;
 use std::sync::{Arc, RwLock};
 
-use crate::entity::{
-    Entity, EventRecord, EventRecordError, BITCODE_PAYLOAD_CODEC, BITCODE_PAYLOAD_CODEC_VERSION,
-};
+use crate::entity::{Entity, EventRecord};
 use crate::outbox::OutboxMessage;
 use crate::read_model::in_memory::apply_read_model_write_plan;
 use crate::read_model::{
@@ -17,9 +15,11 @@ use crate::read_model::{
     ReadModelLoadGraph, ReadModelLoadRequest, ReadModelQueryCapabilities, ReadModelWritePlan,
 };
 use crate::repository::{
-    CommitBatch, GetStream, InboxStore, PreparedEventAppend, ReadModelWritePlanStore,
-    RelationalReadModelQueryStore, RepositoryError, SnapshotStore, SnapshotWrite, StreamIdentity,
-    StreamWrite, TransactionalCommit,
+    reject_duplicate_outbox_messages, reject_duplicate_streams,
+    validate_entity_id_matches_identity, validate_prepared_appends, validate_snapshot_identity,
+    validate_supported_event_codec, CommitBatch, GetStream, InboxStore, PreparedEventAppend,
+    ReadModelWritePlanStore, RelationalReadModelQueryStore, RepositoryError, SnapshotStore,
+    SnapshotWrite, StreamIdentity, StreamWrite, TransactionalCommit,
 };
 use crate::snapshot::{InMemorySnapshotStore, SnapshotRecord};
 
@@ -292,95 +292,10 @@ impl InboxStore for HashMapRepository {
     }
 }
 
-fn reject_duplicate_streams(streams: &[StreamWrite<'_>]) -> Result<(), RepositoryError> {
-    let mut seen = HashSet::with_capacity(streams.len());
-    for stream in streams {
-        let key = stream.identity.storage_key();
-        if !seen.insert(key) {
-            return Err(RepositoryError::DuplicateStreamInBatch {
-                id: stream.identity.to_string(),
-            });
-        }
-    }
-    Ok(())
-}
-
-fn reject_duplicate_outbox_messages(messages: &[OutboxMessage]) -> Result<(), RepositoryError> {
-    let mut seen = HashSet::with_capacity(messages.len());
-    for message in messages {
-        crate::outbox::validate_outbox_message_table_write(message)
-            .map_err(|err| RepositoryError::Model(err.to_string()))?;
-        let id = message.id();
-        if id.trim().is_empty() {
-            return Err(RepositoryError::Model(
-                "outbox message id must not be empty".into(),
-            ));
-        }
-        if message.event_type.trim().is_empty() {
-            return Err(RepositoryError::Model(format!(
-                "outbox message `{id}` event type must not be empty"
-            )));
-        }
-        if !seen.insert(id.to_string()) {
-            return Err(RepositoryError::DuplicateOutboxMessageInBatch { id: id.into() });
-        }
-    }
-    Ok(())
-}
-
-fn validate_entity_id_matches_identity(streams: &[StreamWrite<'_>]) -> Result<(), RepositoryError> {
-    for stream in streams {
-        if stream.entity.id() != stream.identity.aggregate_id() {
-            return Err(RepositoryError::Model(format!(
-                "stream identity `{}` does not match entity id `{}`",
-                stream.identity,
-                stream.entity.id()
-            )));
-        }
-    }
-    Ok(())
-}
-
-fn validate_prepared_appends(appends: &[PreparedEventAppend]) -> Result<(), RepositoryError> {
-    for append in appends {
-        for (offset, event) in append.events.iter().enumerate() {
-            validate_supported_event_codec(event)?;
-            let expected_sequence = append.expected_version + offset as u64 + 1;
-            if event.sequence != expected_sequence {
-                return Err(RepositoryError::Model(format!(
-                    "event `{}` for stream `{}` has sequence {}, expected {}",
-                    event.event_name, append.identity, event.sequence, expected_sequence
-                )));
-            }
-        }
-    }
-    Ok(())
-}
-
-fn validate_supported_event_codec(event: &EventRecord) -> Result<(), RepositoryError> {
-    if event.payload_codec != BITCODE_PAYLOAD_CODEC
-        || event.payload_codec_version != BITCODE_PAYLOAD_CODEC_VERSION
-    {
-        return Err(EventRecordError::unsupported_codec(
-            &event.payload_codec,
-            event.payload_codec_version,
-        )
-        .into());
-    }
-    Ok(())
-}
-
 fn validate_snapshot_write(write: &SnapshotWrite) -> Result<(), RepositoryError> {
     match write {
         SnapshotWrite::Save { identity, record } => validate_snapshot_identity(identity, record),
     }
-}
-
-fn validate_snapshot_identity(
-    identity: &StreamIdentity,
-    record: &SnapshotRecord,
-) -> Result<(), RepositoryError> {
-    record.validate_for_identity(identity)
 }
 
 fn stored_stream_version(events: Option<&Vec<EventRecord>>) -> u64 {
